+++ companion/app/meeting-session-details.tsx
@@ -3,6 +3,7 @@ import { Stack, useLocalSearchParams, useRouter } from "expo-router";
 import { useCallback, useEffect, useState } from "react";
 import { ActivityIndicator, Alert, Platform, View } from "react-native";
 import { AppPressable } from "@/components/AppPressable";
+import { HeaderButtonWrapper } from "@/components/HeaderButtonWrapper";
 import MeetingSessionDetailsScreenComponent from "@/components/screens/MeetingSessionDetailsScreen";
 import { CalComAPIService } from "@/services/calcom";
 import type { ConferencingSession } from "@/services/types/bookings.types";
@@ -34,9 +35,11 @@ export default function MeetingSessionDetails() {
 
   const renderHeaderLeft = useCallback(
     () => (
-      <AppPressable onPress={() => router.back()} className="px-2 py-2">
-        <Ionicons name="close" size={24} color="#007AFF" />
-      </AppPressable>
+      <HeaderButtonWrapper side="left">
+        <AppPressable onPress={() => router.back()} className="px-2 py-2">
+          <Ionicons name="close" size={24} color="#007AFF" />
+        </AppPressable>
+      </HeaderButtonWrapper>
     ),
     [router]
   );
