+++ companion/app/profile-sheet.tsx
@@ -12,6 +12,7 @@ import {
   View,
 } from "react-native";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
+import { HeaderButtonWrapper } from "@/components/HeaderButtonWrapper";
 import { useUserProfile } from "@/hooks";
 import { openInAppBrowser } from "@/utils/browser";
 import { getAvatarUrl } from "@/utils/getAvatarUrl";
@@ -116,9 +117,11 @@ export default function ProfileSheet() {
           },
           headerLeft: () => null,
           headerRight: () => (
-            <TouchableOpacity onPress={handleClose} style={{ padding: 8 }}>
-              <Ionicons name="close" size={24} color="#000" />
-            </TouchableOpacity>
+            <HeaderButtonWrapper side="right">
+              <TouchableOpacity onPress={handleClose} style={{ padding: 8 }}>
+                <Ionicons name="close" size={24} color="#000" />
+              </TouchableOpacity>
+            </HeaderButtonWrapper>
           ),
         }}
       />
