+++ companion/components/booking-list-item/types.ts
@@ -6,11 +6,8 @@ export interface BookingListItemProps {
   isConfirming: boolean;
   isDeclining: boolean;
   onPress: (booking: Booking) => void;
-  onLongPress: (booking: Booking) => void;
   onConfirm: (booking: Booking) => void;
   onReject: (booking: Booking) => void;
-  onActionsPress: (booking: Booking) => void;
-  // Additional action handlers for context menu (iOS)
   onReschedule?: (booking: Booking) => void;
   onEditLocation?: (booking: Booking) => void;
   onAddGuests?: (booking: Booking) => void;
