+++ companion/components/event-type-list-item/EventTypeListItem.ios.tsx
@@ -11,9 +11,7 @@ export const EventTypeListItem = ({
   item,
   index,
   filteredEventTypes,
-  copiedEventTypeId: _copiedEventTypeId,
   handleEventTypePress,
-  handleEventTypeLongPress: _handleEventTypeLongPress,
   handleCopyLink,
   handlePreview,
   onEdit,
