+++ companion/components/event-type-list-item/types.ts
@@ -4,9 +4,7 @@ export interface EventTypeListItemProps {
   item: EventType;
   index: number;
   filteredEventTypes: EventType[];
-  copiedEventTypeId: number | null;
   handleEventTypePress: (item: EventType) => void;
-  handleEventTypeLongPress: (item: EventType) => void;
   handleCopyLink: (item: EventType) => void;
   handlePreview: (item: EventType) => void;
   onEdit?: (item: EventType) => void;
